changed seconds to long

+++ src/main/java/com/getkhaki/api/bff/domain/models/OrganizerStatisticsDm.java
@@ -15,7 +15,7 @@ public class OrganizerStatisticsDm {
     String organizerEmail;
     String organizerFirstName;
     String organizerLastName;
-    int totalMeetings;
+    Integer totalMeetings;
     Double totalCost;
-    Integer totalSeconds;
+    Long totalSeconds;
 }
