+++ src/main/java/com/getkhaki/api/bff/persistence/models/views/DepartmentStatisticsView.java
@@ -7,6 +7,6 @@ public interface DepartmentStatisticsView {
 
     String getDepartmentName();
 
-    Integer getTotalSeconds();
+    Long getTotalSeconds();
 
 }
