+++ src/main/java/com/getkhaki/api/bff/persistence/models/views/OrganizerStatisticsView.java
@@ -9,7 +9,7 @@ public interface OrganizerStatisticsView {
 
     String getOrganizerLastName();
 
-    Integer getTotalSeconds();
+    Long getTotalSeconds();
 
     Double getTotalCost();
 }
