+++ src/main/java/com/getkhaki/api/bff/web/models/OrganizerStatisticsResponseDto.java
@@ -12,6 +12,6 @@ public class OrganizerStatisticsResponseDto {
     String organizerFirstName;
     String organizerLastName;
     Integer totalMeetings;
-    Integer totalSeconds;
+    Long totalSeconds;
     Double totalCost;
 }
